refactor: :truck: Rename heroes_controller to heroes_repository.
